requirements.txt
uvicorn
tensorflow
numpy
firebase-admin
pillow
fastapi
google-cloud-firestore
google-cloud-secret-manager
sqlalchemy
google-auth
cloud-sql-python-connector[pymysql]
pymysql
